+++ code/util/common.h
@@ -38,9 +38,9 @@ inline float saturate( float x )
 
 inline bool is_equal( f32 f0, f32 f1, const f32 eps = FLT_EPSILON )
 {
-    const f32 diff = ::fabs( f0 - f1 );
-    f0 = ::fabs( f0 );
-    f1 = ::fabs( f1 );
+    const f32 diff = ::fabsf( f0 - f1 );
+    f0 = ::fabsf( f0 );
+    f1 = ::fabsf( f1 );
 
     const f32 largest = ( f1 > f0 ) ? f1 : f0;
     return ( diff <= largest * eps ) ? true : false;
